requirements.txt
appdirs==1.4.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
olefile==0.44
packaging==16.8
pandas==0.22.0
pdtools==0.10.2
Pillow==4.0.0
pyparsing==2.1.10
PyYAML==3.12
requests==2.18.4
six==1.10.0
soco==0.13
urllib3==1.22
Werkzeug==0.11.15
